feat(group-lifecycle): read the evidence watermark beside the coverage fraction
A status write needs to prove the evidence it decided on is newer than the
evidence already recorded, and the group's causal tuple cannot say so: RTT
writes never advance it. The coverage computer counted observed edges and
dropped each measurement's version and instant, so it now also reports the
newest measurement it actually counted (product decision 33).

The two are nested rather than flat. GroupFormationReadiness is the shape
GroupFormationView serializes, so widening it would have put an internal
ordering token on the HTTP response without an OpenAPI declaration; a
reading now carries the published triple under `readiness` and the
watermark beside it, which makes leaking it a type error rather than an
oversight. The module and its test are renamed to match the primary export.

Only evidence the reading counted moves the watermark -- a measurement for
an edge outside the plan says nothing about the plan's coverage.
